// task1/exam/src/modules/transaction/dtos/receive-transaction-from-file.dto.ts
import { IsOptional } from 'class-validator';

export class ReceiveTransactionFromFileDto {
  stt: string;

  ngay: string;

  gio: string;

  tram: string;

  truBom: string;

  matHang: string;

  soLuong: string;

  donGia: string;

  thanhTien: string;

  trangThaiThanhToan: string;

  maKhachHang: string;

  tenKhachHang: string;

  loaiKhachHang: string;

  ngayThanhToan: string;

  nhanVien: string;

  bienSoXe: string;

  trangThaiHoaDon: string;
}